refactor: 탈퇴 api 분리
